% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{heatmap}
\alias{heatmap}
\title{This is a data to be included in my package}
\format{An object of class \code{data.frame} with 144 rows and 5 columns.}
\usage{
heatmap
}
\description{
This is a data to be included in my package
}
\references{
\url{https://stat.ethz.ch/R-manual/R-devel/library/stats/html/heatmap.html}
}
\author{
R \email{}
}
\keyword{datasets}
