% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceMovie.R
\docType{package}
\name{spaceMovie}
\alias{spaceMovie}
\alias{spaceMovie-package}
\title{spaceMovie}
\description{
list of palettes from the Space Movie franchise
}
\details{
list of palettes from the Space Movie franchise
}

