% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{SW_palettes}
\alias{SW_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 19.}
\usage{
SW_palettes
}
\description{
Use \code{\link{SW_palette}} to construct palettes of desired length.
}
\keyword{datasets}
